requirements.txt
kafka-python==2.0.2
loguru==0.6.0
